view equipped items menu
