+++ packages/client/src/app/game/ActionMenu/menu-state/inventory-items.ts
@@ -1,126 +1,27 @@
-import { getCurrentMenu, useGameStore } from "@/stores/game-store";
-import {
-  ActionButtonCategory,
-  ActionButtonsByCategory,
-  ActionMenuButtonProperties,
-  ActionMenuState,
-  MenuStateType,
-} from ".";
-import {
-  ConsumableType,
-  Item,
-  ItemPropertiesType,
-  NextOrPrevious,
-  changePage,
-  formatConsumableType,
-} from "@speed-dungeon/common";
-import { setAlert } from "@/app/components/alerts";
-import { ConsideringItemMenuState } from "./considering-item";
-import { ACTION_MENU_PAGE_SIZE } from "..";
+import { useGameStore } from "@/stores/game-store";
+import { ActionButtonCategory, ActionMenuButtonProperties, MenuStateType } from ".";
 import { immerable } from "immer";
+import { ItemsMenuState } from "./items";
+import { EquippedItemsMenuState } from "./equipped-items";
 
-export class InventoryItemsMenuState implements ActionMenuState {
+export class InventoryItemsMenuState extends ItemsMenuState {
   [immerable] = true;
-  type = MenuStateType.InventoryItems;
   page = 1;
   numPages = 1;
-  constructor() {}
-  getButtonProperties(): ActionButtonsByCategory {
-    const toReturn = new ActionButtonsByCategory();
-
-    const closeInventory = new ActionMenuButtonProperties("Close Inventory", () => {
+  constructor() {
+    const viewEquipmentButton = new ActionMenuButtonProperties("View Equipped", () => {
       useGameStore.getState().mutateState((state) => {
-        state.stackedMenuStates.pop();
+        state.stackedMenuStates.push(new EquippedItemsMenuState());
       });
     });
-    closeInventory.dedicatedKeys = ["KeyI", "KeyS", "Escape"];
-    toReturn[ActionButtonCategory.Top].push(closeInventory);
+    viewEquipmentButton.dedicatedKeys = ["KeyF"];
 
-    const toggleViewEquippedItems = new ActionMenuButtonProperties("Show Equipped", () => {
-      // @TODO
-    });
-
-    toggleViewEquippedItems.dedicatedKeys = ["KeyF"];
-    toReturn[ActionButtonCategory.Top].push(toggleViewEquippedItems);
-
-    let focusedCharacterResult = useGameStore.getState().getFocusedCharacter();
-    if (focusedCharacterResult instanceof Error) {
-      setAlert(focusedCharacterResult.message);
-      return toReturn;
-    }
-    const { combatantProperties } = focusedCharacterResult;
-
-    const equipment: Item[] = [];
-    const consumablesByType: Partial<Record<ConsumableType, Item[]>> = {};
-
-    for (const item of Object.values(combatantProperties.inventory.items)) {
-      switch (item.itemProperties.type) {
-        case ItemPropertiesType.Equipment:
-          equipment.push(item);
-          break;
-        case ItemPropertiesType.Consumable:
-          const { consumableType } = item.itemProperties.consumableProperties;
-          if (!consumablesByType[consumableType]) consumablesByType[consumableType] = [item];
-          else consumablesByType[consumableType]!.push(item);
+    super(
+      MenuStateType.InventoryItems,
+      { text: "Close Inventory", hotkeys: ["KeyI", "KeyS"] },
+      {
+        [ActionButtonCategory.Top]: [viewEquipmentButton],
       }
-    }
-
-    for (const [consumableTypeKey, consumables] of Object.entries(consumablesByType)) {
-      const firstConsumableOfThisType = consumables[0];
-      if (!firstConsumableOfThisType) continue;
-      let consumableName = formatConsumableType(parseInt(consumableTypeKey));
-      if (consumables.length > 1) consumableName += ` (${consumables.length})`;
-
-      const button = new ActionMenuButtonProperties(consumableName, () => {
-        useGameStore.getState().mutateState((state) => {
-          state.stackedMenuStates.push(new ConsideringItemMenuState(firstConsumableOfThisType));
-        });
-      });
-      toReturn[ActionButtonCategory.Numbered].push(button);
-    }
-
-    for (const item of equipment) {
-      const button = new ActionMenuButtonProperties(item.entityProperties.name, () => {
-        useGameStore.getState().mutateState((state) => {
-          state.stackedMenuStates.push(new ConsideringItemMenuState(item));
-        });
-      });
-      toReturn[ActionButtonCategory.Numbered].push(button);
-    }
-
-    const numPages = Math.ceil(
-      toReturn[ActionButtonCategory.Numbered].length / ACTION_MENU_PAGE_SIZE
     );
-
-    useGameStore.getState().mutateState((state) => {
-      getCurrentMenu(state).numPages = numPages;
-    });
-
-    if (numPages > 1) {
-      const previousPageButton = new ActionMenuButtonProperties("Previous", () => {
-        useGameStore.getState().mutateState((state) => {
-          const newPage = changePage(this.page, numPages, NextOrPrevious.Previous);
-          getCurrentMenu(state).page = newPage;
-        });
-      });
-      previousPageButton.dedicatedKeys = ["KeyW", "ArrowLeft"];
-      toReturn[ActionButtonCategory.Bottom].push(previousPageButton);
-
-      const nextPageButton = new ActionMenuButtonProperties("Next", () => {
-        useGameStore.getState().mutateState((state) => {
-          const newPage = changePage(this.page, numPages, NextOrPrevious.Next);
-          getCurrentMenu(state).page = newPage;
-        });
-      });
-      nextPageButton.dedicatedKeys = ["KeyE", "ArrowRight"];
-      toReturn[ActionButtonCategory.Bottom].push(nextPageButton);
-    }
-
-    toReturn[ActionButtonCategory.Numbered] = toReturn[ActionButtonCategory.Numbered].slice(
-      (this.page - 1) * ACTION_MENU_PAGE_SIZE,
-      (this.page - 1) * ACTION_MENU_PAGE_SIZE + ACTION_MENU_PAGE_SIZE
-    );
-
-    return toReturn;
   }
 }
